feat: 연관관계 메서드 및 비즈니스 로직과 equals,hashCode 작성
